+++ fridayboot-admin/src/main/java/org/knifez/fridaybootadmin/entity/AppUserRole.java
@@ -29,13 +29,13 @@ public class AppUserRole extends BaseAuditEntity implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Long id;
+    private Integer id;
 
     @Schema(title = "用户id")
-    private Long userId;
+    private Integer userId;
 
     @Schema(title = "角色id")
-    private Long roleId;
+    private Integer roleId;
 
 
 }
